Refactor: move helper classes to own files
