+++ compare_hashes/compare_hashes.cpp
@@ -12,8 +12,10 @@
 
 #include "boost/filesystem/operations.hpp"
 #include "boost/filesystem/path.hpp"
-
 #include "nlohmann/json.hpp"
+
+#include "submission.h"
+#include "hash_location.h"
 
 
 // ===================================================================================
@@ -26,16 +28,6 @@ typedef std::string hash;
 // ===================================================================================
 // helper classes
 
-// represents the location of a hash within
-// a unique student and version pair
-struct HashLocation {
-  HashLocation(const std::string &s, int v, location_in_submission l, const std::string &sg) : student(s), version(v), location(l), source_gradeable(sg) {}
-  std::string student;
-  int version;
-  location_in_submission location;
-  std::string source_gradeable;
-};
-
 
 // represents an element in a ranking of students by percent match
 struct StudentRanking {
@@ -47,67 +39,8 @@ struct StudentRanking {
 };
 
 
-// represents a unique student-version pair, all its
-// hashes, and other submissions with those hashes
-class Submission {
-public:
-  // CONSTRUCTOR
-  Submission(const std::string &s, int v) : student_(s), version_(v) {}
-
-  // GETTERS
-  const std::string & student() const { return student_; }
-  int version() const { return version_; }
-
-  const std::map<location_in_submission, std::set<HashLocation> >& getSuspiciousMatches() const {
-    return suspicious_matches;
-  }
-  const std::set<location_in_submission>& getCommonMatches() const { return common_matches; }
-  const std::set<location_in_submission>& getProvidedMatches() const { return provided_matches; }
-  const std::unordered_map<std::string, std::unordered_map<std::string, std::unordered_map<int, std::unordered_set<hash>>>>& getStudentsMatched() const {
-    return students_matched;
-  }
-  float getPercentage() const {
-    return (100.0 * (suspicious_matches.size())) / hashes.size();
-  }
-
-  // MODIFIERS
-  void addHash(const hash &h, location_in_submission l) { hashes.push_back(make_pair(h, l)); }
-  const std::vector<std::pair<hash, location_in_submission>> & getHashes() const { return hashes; }
-
-  void addSuspiciousMatch(location_in_submission location, const HashLocation &matching_location, hash matched_hash) {
-    // save the found match
-    suspicious_matches[location].insert(matching_location);
-    // update the students_matched container
-    students_matched[matching_location.source_gradeable][matching_location.student][matching_location.version].insert(matched_hash);
-  }
-
-  void addCommonMatch(location_in_submission location) { common_matches.insert(location); }
-  void addProvidedMatch(location_in_submission location) { provided_matches.insert(location); }
-
-private:
-  std::string student_;
-  int version_;
-  std::vector<std::pair<hash, location_in_submission> > hashes;
-  std::map<location_in_submission, std::set<HashLocation> > suspicious_matches;
-  std::set<location_in_submission> common_matches;
-  std::set<location_in_submission> provided_matches;
-
-  // a container to keep track of all the students this submission
-  // matched and the number of matching hashes per submission
-  // <source_gradeable, <username, <version, <hashes>> > >
-  std::unordered_map<std::string, std::unordered_map<std::string, std::unordered_map<int, std::unordered_set<hash>>>> students_matched;
-};
-
-
 // ===================================================================================
 // helper functions
-
-bool operator < (const HashLocation &hl1, const HashLocation &hl2) {
-  return hl1.student > hl2.student ||
-         (hl1.student == hl2.student && hl1.version < hl2.version) ||
-         (hl1.student == hl2.student && hl1.version == hl2.version && hl1.location < hl2.location);
-}
-
 
 // ensures that all of the regions in the two parameters are adjacent
 bool matchingPositionsAreAdjacent(const nlohmann::json &first, const nlohmann::json &second) {
